% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{apath_dependency}
\alias{apath_dependency}
\title{Functions}
\usage{
apath_dependency()
}
\description{
\code{apath} dependency.
}
\examples{
library(aframer)

a_scene(
  a_dependency(),
  apath_dependency()
)

}
